Move ConcurrentMap from ComponentMapping to it's own file for reuseability
extracts the ConcurrentMap created by Sean to it’s own file so it can
be reused in other parts of the Go WallarooAPI when needed.
